feat(api): move `Mutant` interface
Move the `Mutant` interface from from `api/mutant` to `api/core`.
It is now also used in the new checker api.
