+++ packages/api/core.ts
@@ -2,6 +2,7 @@ export { default as File } from './src/core/File';
 export { default as Position } from './src/core/Position';
 export { default as Location } from './src/core/Location';
 export { default as Range } from './src/core/Range';
+export { default as Mutant } from './src/core/Mutant';
 export * from './src-generated/stryker-core';
 export * from './src/core/ReportTypes';
 export * from './src/core/StrykerOptionsSchema';
