+++ packages/api/mutant.ts
@@ -1,2 +1,1 @@
-export { default as Mutant } from './src/mutant/Mutant';
 export { default as Mutator } from './src/mutant/Mutator';
