+++ packages/typescript/src/mutator/NodeMutator.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
 
-import { Mutant } from '@stryker-mutator/api/mutant';
+import { Mutant } from '@stryker-mutator/api/core';
 import * as ts from 'typescript';
 
 export interface NodeReplacement {
